requirements.txt
Django==4.2.9
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==23.5
Pillow==10.2.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-allauth==0.60.1
django-cors-headers==4.3.1
django-extensions==3.2.3
whitenoise==6.6.0
python-decouple==3.8
dj-database-url==2.1.0
django-ckeditor-5==0.2.20
django-meta==2.3.0
django-taggit==5.0.1
djangorestframework-simplejwt==5.3.1
django-import-export==3.3.6
drf-spectacular==0.27.2
gunicorn==21.2.0
